Fix bug where data would be pasted from wrong cell range.
Formerly, when cut-n-pasting the data_received callback would paste_from the
selected range.  This was ok when pasting to another application, but when
cut-n-pasting within the same sheet, the selected range must have changed.

This change solves the problem by saving the selected range and attaching
it to the clipboard object.