ssw-sheet-axis.c: add check to out-of-sight processing (bug #55357)
ensure_visible_widgets does an out-of-sight processing which is used
for fast scrolling. This process does not work correctly if there are
currently no active widgets. This is the case when variables are
deleted. The problem is that the estimated_widget_size returns 0 when
no widgets are in the window. This fixes:

https://savannah.gnu.org/bugs/?55357

where deleting a variable takes a couple of minutes.