remove force_update callback to prevent stall (#55357)
The force_update callback was required to update
the variable view window in case the data window
is not realized yet. However this results in a
stall when variables are deleted. If only the
sheet is updated that is currently visible, then
this is not required and the stall is prevented.
The items_changed_cb is sufficient but pspp called
this on a widget from the datasheet which is not
realized directly after startup when only the variable
view is seen. That prevents the correct redraw.

Closes: #55357